fix: Resolve build error by separating server-only code
Move the updateLocalBlogManifest function to a separate .server.ts file
to prevent Next.js from trying to bundle fs/promises in client-side code.
This fixes the build error 'Module not found: Can't resolve fs/promises'.
